SONARIAC-813 Rule S6364: ARM Defining a short backup retention duration is security-sensitive (JSON) - Microsoft.RecoveryServices/vaults

+++ iac-extensions/arm/src/main/java/org/sonar/iac/arm/tree/api/NumericLiteral.java
@@ -20,5 +20,5 @@
 package org.sonar.iac.arm.tree.api;
 
 public interface NumericLiteral extends Expression {
-  float value();
+  double value();
 }
